requirements.txt
redshift_connector
apache-airflow[amazon]
apache-airflow-providers-amazon
yfinance
pandas